line_bot_sample1/requirements.txt
boto3==1.7.70
botocore==1.10.70
certifi==2018.1.18
chardet==3.0.4
Click==7.0
cycler==0.10.0
docutils==0.14
Flask==1.0.2
future==0.16.0
futures==3.1.1
gunicorn==19.9.0
httpie==0.9.9
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
line-bot-sdk==1.8.0
MarkupSafe==1.0
matplotlib==2.0.2
msgpack-python==0.5.6
nose==1.3.7
psycopg2
numpy==1.14.5
protobuf==3.6.0
Pygments==2.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
requests==2.18.4
s3transfer==0.1.13
scipy==1.1.0
freeze==1.0.10
six==1.10.0
subprocess32
urllib3==1.22
Werkzeug==0.14.1